Move USB data types to a separate module.
